src/MetaSeg/requirements.txt
Cython==0.29.13
h5py==2.10.0
matplotlib==3.0.3
numpy==1.17.2
pandas==0.24.2
Pillow==6.1.0
scikit-learn==0.21.3
